Separate the terms of service label to a reusable view
